pythonScript/pythonProject/requirements.txt
paho-mqtt
requests

